refactored fftshift and ifftshift code
