+++ tensorflow/python/ops/signal/signal.py
@@ -62,6 +62,6 @@ from tensorflow.python.ops.signal.spectral_ops import inverse_stft_window_fn
 from tensorflow.python.ops.signal.spectral_ops import stft
 from tensorflow.python.ops.signal.window_ops import hamming_window
 from tensorflow.python.ops.signal.window_ops import hann_window
-from tensorflow.python.ops.signal.helper import fftshift
-from tensorflow.python.ops.signal.helper import ifftshift
+from tensorflow.python.ops.signal.fft_ops import fftshift
+from tensorflow.python.ops.signal.fft_ops import ifftshift
 # pylint: enable=unused-import
